requirements/base.txt
bootstrap-admin==0.4.0
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.37.1
django-crispy-forms==1.7.2
django-dotenv==1.4.2
django-filter==2.0.0
djangorestframework==3.8.2
idna==2.7
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.2.0
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23